+++ tests/acceptance/bashunit_init_test.sh
@@ -15,7 +15,7 @@ function tear_down() {
 
 function test_bashunit_init_creates_structure() {
   pushd "$TMP_DIR" >/dev/null
-  "$BASHUNIT_PATH" init >/tmp/init.log
+  "$BASHUNIT_PATH" init >"$TMP_DIR/init.log"
   assert_file_exists "tests/example_test.sh"
   assert_file_exists "tests/bootstrap.sh"
   popd >/dev/null
@@ -23,7 +23,7 @@ function test_bashunit_init_creates_structure() {
 
 function test_bashunit_init_custom_directory() {
   pushd "$TMP_DIR" >/dev/null
-  "$BASHUNIT_PATH" init custom >/tmp/init.log
+  "$BASHUNIT_PATH" init custom >"$TMP_DIR/init.log"
   assert_file_exists "custom/example_test.sh"
   assert_file_exists "custom/bootstrap.sh"
   popd >/dev/null
@@ -31,7 +31,7 @@ function test_bashunit_init_custom_directory() {
 
 function test_bashunit_init_creates_github_workflow() {
   pushd "$TMP_DIR" >/dev/null
-  "$BASHUNIT_PATH" init >/tmp/init.log
+  "$BASHUNIT_PATH" init >"$TMP_DIR/init.log"
   assert_file_exists ".github/workflows/tests.yml"
   assert_file_contains ".github/workflows/tests.yml" "TypedDevs/bashunit@"
   popd >/dev/null
@@ -41,7 +41,7 @@ function test_bashunit_init_does_not_overwrite_existing_workflow() {
   pushd "$TMP_DIR" >/dev/null
   mkdir -p ".github/workflows"
   echo "custom-workflow" >".github/workflows/tests.yml"
-  "$BASHUNIT_PATH" init >/tmp/init.log
+  "$BASHUNIT_PATH" init >"$TMP_DIR/init.log"
   assert_file_contains ".github/workflows/tests.yml" "custom-workflow"
   popd >/dev/null
 }
@@ -51,7 +51,7 @@ function test_bashunit_init_updates_env() {
 
   pushd "$TMP_DIR" >/dev/null
   echo "BASHUNIT_BOOTSTRAP=old/bootstrap.sh" >.env
-  "$BASHUNIT_PATH" init custom >/tmp/init.log
+  "$BASHUNIT_PATH" init custom >"$TMP_DIR/init.log"
   assert_file_exists "custom/example_test.sh"
   assert_file_exists "custom/bootstrap.sh"
   assert_file_contains .env "#BASHUNIT_BOOTSTRAP=old/bootstrap.sh"