+++ tests/acceptance/bashunit_syntax_error_test.sh
@@ -10,7 +10,7 @@ function test_bashunit_when_test_file_has_syntax_error() {
 
   local actual_raw
   set +e
-  actual_raw="$(LC_ALL=C LANG=C ./bashunit \
+  actual_raw="$(./bashunit \
     --no-parallel --detailed --env "$TEST_ENV_FILE" "$test_file" 2>&1)"
   set -e
 
@@ -19,6 +19,6 @@ function test_bashunit_when_test_file_has_syntax_error() {
 
   assert_contains "failed" "$actual"
   assert_contains "Error" "$actual"
-  assert_general_error "$(LC_ALL=C LANG=C ./bashunit \
+  assert_general_error "$(./bashunit \
     --no-parallel --env "$TEST_ENV_FILE" "$test_file" 2>&1)"
 }